fixed integration tests for email infrastructure
