+++ GameSync.Api.IntegrationTests/GameSync.Api.IntegrationTests.csproj
@@ -17,6 +17,7 @@
     <PackageReference Include="NSubstitute" Version="5.1.0" />
     <PackageReference Include="xunit" Version="2.5.3" />
     <PackageReference Include="xunit.runner.visualstudio" Version="2.5.3" />
+    <PackageReference Include="Pose" Version="1.2.1" />
   </ItemGroup>
 
   <ItemGroup>
